update the excel logic to be generic based on the user cards

+++ src/services/expense.service.ts
@@ -1,10 +1,12 @@
+import { IUserDocument } from './../models/user.model';
 import { IExpenseDocument, ExpenseModel } from '../models/expense.model';
 import { DocumentDefinition, FilterQuery } from 'mongoose';
 import excelToJson from 'convert-excel-to-json';
 import fs from 'fs';
+import { UserModel } from '../models/user.model';
 
-const setDataToBeSavedInDb = (expenses: any[], userId: string): IExpenseDocument[] => {
-  const filtered = expenses.filter((item) => item.cardNumber === '9362');
+const setDataToBeSavedInDb = (expenses: any[], userId: string, userCards: string[]): IExpenseDocument[] => {
+  const filtered = expenses.filter((item) => userCards.find((card) => card === item.cardNumber));
   const expensesToSave = filtered.map((item) => {
     return {
       ...item,
@@ -52,10 +54,22 @@ const expenseService = {
   expenseFromExcel: async (userId: string, excel: any): Promise<IExpenseDocument[]> => {
     try {
       const json = convertToJson(excel.path);
-      const expensesToSave = setDataToBeSavedInDb(json.sheet1, userId);
-      const expenses = await ExpenseModel.insertMany(expensesToSave);
+      const user = (await UserModel.findById(userId)) as IUserDocument;
+      const userCards = user.userCards;
+      if (!userCards) throw new Error('user should have at least one card');
+
+      let expensesList: IExpenseDocument[] = [];
+
+      for (const [key, sheet] of Object.entries(json)) {
+        const expensesToSave = setDataToBeSavedInDb(sheet, userId, userCards);
+        expensesList = expensesList.concat(expensesToSave);
+      }
+      console.log('expensesList', expensesList);
+
+      // const expenses = await ExpenseModel.insertMany(expensesList);
       fs.unlinkSync(excel.path);
-      return expenses;
+      // return expenses;
+      return [];
     } catch (error) {
       throw error;
     }
